improve specs
